Move container definition into image module
